requirements.txt
brotlipy==0.7.0
django-admin-sortable2==1.0.4
django-tinymce==3.6.1
environs==9.5.0
Pillow==9.5.0
django==3.2.18
